fix capping rule when the meta only contains impression

+++ spark-nrt/src/main/scala/com/ebay/traffic/chocolate/sparknrt/capping/rules/IPCappingRule.scala
@@ -12,6 +12,7 @@ import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.functions.split
 import org.apache.spark.sql.functions.count
 import org.apache.spark.sql.functions.lit
+import org.apache.spark.sql.functions.when
 import org.apache.spark.sql.functions.sum
 import org.slf4j.LoggerFactory
 
@@ -51,59 +52,67 @@ class IPCappingRule(params: Parameter, bit: Long, dateFiles: DateFiles, cappingR
 
     // filter click only, count ip and save to tmp file
     var dfIP = cappingRuleJobObj.readFilesAsDFEx(dateFiles.files).filter($"channel_action" === "CLICK")
-    val timestamp = dfIP.select($"timestamp").first().getLong(0)
-
-    dfIP = dfIP.select(split($"request_headers", "X-eBay-Client-IP: ")(1).alias("tmpIP"))
-      .select(split($"tmpIP", """\|""")(0).alias("IP"))
-      .groupBy($"IP").agg(count(lit(1)).alias("count"))
-      .drop($"request_headers")
-      .drop($"tmpIP")
-
-    // reduce the number of ip count file to 1
-    dfIP = dfIP.repartition(1)
-    val tempPath = baseTempDir + dateFiles.date
-    cappingRuleJobObj.saveDFToFiles(dfIP, tempPath)
-
-    // rename file name to include timestamp
-    val fileStatus = fs.listStatus(new Path(tempPath))
-    val src = fileStatus.filter(status => status.getPath.getName != "_SUCCESS")(0).getPath
-    val target = new Path(tempPath, s"part-${timestamp}.snappy.parquet")
-    fs.rename(src, target)
-
-    // IP rule
-    var df = cappingRuleJobObj.readFilesAsDFEx(dateFiles.files)
-      .withColumn("tmpIP", split($"request_headers", "X-eBay-Client-IP: ")(1))
-      .withColumn("IP_1", split($"tmpIP", """\|""")(0))
-      .drop($"tmpIP")
-
-    val ipCountTempPathToday = baseTempDir + dateFiles.date
-    val ipCountPathToday = baseDir + dateFiles.date
-    val cal = Calendar.getInstance
-    cal.setTime(sdf.parse((dateFiles.date.asInstanceOf[String]).substring(DATE_COL.length + 1)))
-    cal.add(Calendar.DATE, -1)
-    val dateBefore1Day = cal.getTime
-    val ipCountTempPathYesterday = baseTempDir + DATE_COL + "=" + sdf.format(dateBefore1Day)
-    val ipCountPathYesterday = baseDir + DATE_COL + "=" + sdf.format(dateBefore1Day)
-    var ipCountPath: List[String] = List()
-    ipCountPath = ipCountPath :+ ipCountTempPathToday
-    if (fs.exists(new Path(ipCountPathToday))) {
-      ipCountPath = ipCountPath :+ ipCountPathToday
+    val head = dfIP.take(1)
+    if(head.length == 0) {
+      var df = cappingRuleJobObj.readFilesAsDFEx(dateFiles.files).withColumn("capping", lit(0l))
+      df
     }
-    // read only 24 hours data
-    if (fs.exists(new Path(ipCountPathYesterday))) {
-      val fileStatus = fs.listStatus(new Path(ipCountPathYesterday))
-        .filter(status => String.valueOf(status.getPath.getName.substring(DATE_COL.length + 1, status.getPath.getName.indexOf("."))) >= (timestamp - 86400000).toString)
-        .map(status => ipCountPath = ipCountPath :+ status.getPath.toString)
+    else {
+      val firstRow = head(0)
+      val timestamp = firstRow.getLong(firstRow.fieldIndex("timestamp"));
+
+      dfIP = dfIP.select(split($"request_headers", "X-eBay-Client-IP: ")(1).alias("tmpIP"))
+        .select(split($"tmpIP", """\|""")(0).alias("IP"))
+        .groupBy($"IP").agg(count(lit(1)).alias("count"))
+        .drop($"request_headers")
+        .drop($"tmpIP")
+
+      // reduce the number of ip count file to 1
+      dfIP = dfIP.repartition(1)
+      val tempPath = baseTempDir + dateFiles.date
+      cappingRuleJobObj.saveDFToFiles(dfIP, tempPath)
+
+      // rename file name to include timestamp
+      val fileStatus = fs.listStatus(new Path(tempPath))
+      val src = fileStatus.filter(status => status.getPath.getName != "_SUCCESS")(0).getPath
+      val target = new Path(tempPath, s"part-${timestamp}.snappy.parquet")
+      fs.rename(src, target)
+
+      // IP rule
+      var df = cappingRuleJobObj.readFilesAsDFEx(dateFiles.files)
+        .withColumn("tmpIP", split($"request_headers", "X-eBay-Client-IP: ")(1))
+        .withColumn("IP_1", when($"channel_action" === "CLICK", split($"tmpIP", """\|""")(0)).otherwise("NA"))
+        .drop($"tmpIP")
+
+      val ipCountTempPathToday = baseTempDir + dateFiles.date
+      val ipCountPathToday = baseDir + dateFiles.date
+      val cal = Calendar.getInstance
+      cal.setTime(sdf.parse((dateFiles.date.asInstanceOf[String]).substring(DATE_COL.length + 1)))
+      cal.add(Calendar.DATE, -1)
+      val dateBefore1Day = cal.getTime
+      val ipCountTempPathYesterday = baseTempDir + DATE_COL + "=" + sdf.format(dateBefore1Day)
+      val ipCountPathYesterday = baseDir + DATE_COL + "=" + sdf.format(dateBefore1Day)
+      var ipCountPath: List[String] = List()
+      ipCountPath = ipCountPath :+ ipCountTempPathToday
+      if (fs.exists(new Path(ipCountPathToday))) {
+        ipCountPath = ipCountPath :+ ipCountPathToday
+      }
+      // read only 24 hours data
+      if (fs.exists(new Path(ipCountPathYesterday))) {
+        val fileStatus = fs.listStatus(new Path(ipCountPathYesterday))
+          .filter(status => String.valueOf(status.getPath.getName.substring(DATE_COL.length + 1, status.getPath.getName.indexOf("."))) >= (timestamp - 86400000).toString)
+          .map(status => ipCountPath = ipCountPath :+ status.getPath.toString)
+      }
+
+      dfIP = cappingRuleJobObj.readFilesAsDFEx(ipCountPath.toArray).groupBy($"IP").agg(sum($"count") as "amnt").filter($"amnt" >= params.ipThreshold)
+        .withColumn("capping", lit(cappingBit)).drop($"count").drop($"amnt")
+
+      df = df.join(dfIP, $"IP_1" === $"IP", "left_outer")
+        .select(df.col("*"), $"capping")
+        .drop("IP_1")
+        .drop("IP")
+      df
     }
-
-    dfIP = cappingRuleJobObj.readFilesAsDFEx(ipCountPath.toArray).groupBy($"IP").agg(sum($"count") as "amnt").filter($"amnt" >= params.ipThreshold)
-      .withColumn("capping", lit(cappingBit)).drop($"count").drop($"amnt")
-
-    df = df.join(dfIP, $"IP_1" === $"IP", "left_outer")
-      .select(df.col("*"), $"capping")
-      .drop("IP_1")
-      .drop("IP")
-    df
   }
 
   override def postTest() = {
@@ -112,14 +121,15 @@ class IPCappingRule(params: Parameter, bit: Long, dateFiles: DateFiles, cappingR
     if (!fs.exists(dateOutputPath)) {
       fs.mkdirs(dateOutputPath)
     }
+    if(fs.exists(new Path(baseTempDir + dateFiles.date))) {
+      val fileStatus = fs.listStatus(new Path(baseTempDir + dateFiles.date))
+      val src = fileStatus.filter(status => status.getPath.getName != "_SUCCESS").toList(0).getPath
+      val fileName = src.getName
+      val dest = new Path(dateOutputPath, fileName)
+      fs.rename(src, dest)
 
-    val fileStatus = fs.listStatus(new Path(baseTempDir + dateFiles.date))
-    val src = fileStatus.filter(status => status.getPath.getName != "_SUCCESS").toList(0).getPath
-    val fileName = src.getName
-    val dest = new Path(dateOutputPath, fileName)
-    fs.rename(src, dest)
-
-    // delete the tmp dir
-    fs.delete(new Path(baseTempDir), true)
+      // delete the tmp dir
+      fs.delete(new Path(baseTempDir), true)
+    }
   }
 }
